requirements.txt
awesome-slugify==1.6.5
click==7.1.2
prometheus-client==0.8.0
pymodbus==2.4.0
